MC-34499: [Integration Tests] 85 Failing in CE on 2.3-develop

+++ dev/tests/integration/testsuite/Magento/Catalog/_files/category_with_different_price_products_rollback.php
@@ -38,9 +38,12 @@ try {
         ->addAttributeToFilter(CategoryInterface::KEY_NAME, 'Category 999')
         ->setPageSize(1)
         ->getFirstItem();
-    $categoryRepository->delete($category);
+    if ($category->getId()) {
+        $categoryRepository->delete($category);
+    }
 } catch (NoSuchEntityException $e) {
     //Already deleted.
 }
+
 $registry->unregister('isSecureArea');
 $registry->register('isSecureArea', false);
